use Application ID as identifier for included apps
The Application ID is a guaranteed, global, unique identifier used in
Android, the APK, Google Play, F-Droid, etc.  It can be easily parsed out
of APKs so it is easy to use in automation.  It is also used in the
standard file naming scheme of fdroid repos.  This changes this repo to use
it as each module name and to look for files based on that naming. This has
to happen in coordination with other changes, so this is just getting that
ball rolling.

Change-Id: I5b12d546c6cf4a46fa08654b56f3c136b8f74f44